+++ crates/teloxide-core/src/types/poll_answer.rs
@@ -21,6 +21,11 @@ pub struct PollAnswer {
     ///
     /// May be empty if the user retracted their vote.
     pub option_ids: Vec<u8>,
+
+    /// Persistent identifiers of the chosen answer options. May be empty if the
+    /// vote was retracted.
+    #[serde(default)]
+    pub option_persistent_ids: Vec<String>,
 }
 
 /// These fields `chat` and `user` from the original [`PollAnswer`] should be