+++ crates/teloxide/src/utils/markdown.rs
@@ -308,6 +308,7 @@ mod tests {
             added_to_attachment_menu: false,
             has_topics_enabled: false,
             allows_users_to_create_topics: false,
+            can_manage_bots: false,
         };
         assert_eq!(user_mention_or_link(&user_with_username), "@abcd");
         // Telegram allows underscores in usernames
@@ -326,6 +327,7 @@ mod tests {
             added_to_attachment_menu: false,
             has_topics_enabled: false,
             allows_users_to_create_topics: false,
+            can_manage_bots: false,
         };
         assert_eq!(user_mention_or_link(&user_without_username), "[Name](tg://user/?id=123456789)")
     }